refactor(esmodule): move bootstrap code out of index.js
The bootstrap code has hard-references to all the components,
and having that in index.js causes `import { Modal } from 'carbon-components'` code bring in all components.
