+++ package.json
@@ -45,6 +45,7 @@
     "babel-plugin-dev-expression": "^0.2.1",
     "babel-plugin-external-helpers": "^6.22.0",
     "babel-plugin-istanbul": "^3.0.0",
+    "babel-plugin-rewire": "^1.1.0",
     "babel-plugin-transform-class-properties": "^6.22.0",
     "babel-plugin-transform-es2015-modules-umd": "^6.22.0",
     "babel-plugin-transform-runtime": "^6.22.0",
